+++ sospice/util/fov.py
@@ -1,13 +1,15 @@
 from dataclasses import dataclass
-from tempfile import TemporaryDirectory
-from pathlib import Path
+from platformdirs import user_data_path
 
 import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
 
 from astropy.coordinates import SkyCoord
-from sunpy.map import Map
+import astropy.units as u
+
+from sunpy.map import Map, make_fitswcs_header
+from sunpy.coordinates import frames
 from sunpy.coordinates.sun import carrington_rotation_number
 from sunpy.net import Fido, attrs
 import sunpy_soar  # noqa: F401
@@ -102,16 +104,21 @@ class FovBackground:
         carrington_rotation = int(np.floor(carrington_rotation_number(self.time)))
         url = f"http://jsoc.stanford.edu/data/hmi/synoptic/hmi.Synoptic_Mr.{carrington_rotation}.fits"
         syn_map = Map(url)
-        fig = plt.figure(figsize=(18, 8))
+        fig = plt.figure(figsize=(19, 9.5))
         ax = fig.add_subplot(projection=syn_map)
         syn_map.plot(
             axes=ax, title=f"SDO/HMI synoptic map for CR {carrington_rotation}"
         )
         return fig, ax
 
-    def plot_EUI_FSI(self):
+    def plot_EUI_FSI(self, delta_t=pd.Timedelta(hours=1)):
         """
         Plot Solar Orbiter/EUI/FSI map, intended as a background for plotting SPICE FOVs
+
+        Parameters
+        ----------
+        delta_t: pandas.Timedelta
+            Half-width of time interval in which to look to EUI data
 
         Return
         ------
@@ -120,7 +127,6 @@ class FovBackground:
         matplotlib.axes.Axes
             Axes (with relevant projection)
         """
-        delta_t = pd.Timedelta(minutes=30)
         results_fsi = Fido.search(
             attrs.Time(self.time - delta_t, self.time + delta_t),
             attrs.soar.Product("eui-fsi174-image"),
@@ -130,19 +136,18 @@ class FovBackground:
         str_n_found = f"for {self.time} ± {delta_t}"
         if n_found == 0:
             raise RuntimeError("No file found " + str_n_found)
+            # TODO rather revert to blank map
         else:
             print(f"{n_found} files found for {str_n_found}")
         delay = pd.Series(results_fsi[0]["Start time"]).apply(pd.Timestamp) - self.time
         i_closest = abs(delay).argmin()
-        with TemporaryDirectory(prefix="tmp-fsi-") as tmp_dir:
-            fsi_file = Fido.fetch(
-                results_fsi[0][i_closest], path=Path(tmp_dir) / "{file}"
-            )
-            fsi_filename = fsi_file[0]
-            fig = plt.figure(figsize=(10, 10))
-            fsi_map = Map(fsi_filename)
-            ax = fig.add_subplot(projection=fsi_map)
-            fsi_map.plot(axes=ax)
+        path = user_data_path(appname="sospice", ensure_exists=True)
+        fsi_file = Fido.fetch(results_fsi[0][i_closest], path=path / "{file}")
+        fsi_filename = fsi_file[0]
+        fig = plt.figure(figsize=(10, 10))
+        fsi_map = Map(fsi_filename)
+        ax = fig.add_subplot(projection=fsi_map)
+        fsi_map.plot(axes=ax)
         return fig, ax
 
     def plot_blank_helioprojective(self):
@@ -156,11 +161,33 @@ class FovBackground:
         matplotlib.axes.Axes
             Axes (with relevant projection)
         """
-        raise NotImplementedError("Blank map background not implemented yet")
-        return
-        # return fig, ax
+        data = np.full((10, 10), np.nan)
 
-    def plot_map(self, show=True, save=None):
+        obs_heligraphic_sth = self.observer
+        obstime = self.observer.obstime
+
+        skycoord = SkyCoord(
+            0 * u.arcsec,
+            0 * u.arcsec,
+            obstime=obstime,
+            observer=obs_heligraphic_sth,
+            frame=frames.Helioprojective,
+        )
+        header = make_fitswcs_header(
+            data, skycoord, scale=[1000, 1000] * u.arcsec / u.pixel
+        )
+        blank_map = Map(data, header)
+
+        # fig = plt.figure()
+        fig = plt.figure(figsize=(20, 10))
+        ax = fig.add_subplot(projection=blank_map)
+        blank_map.plot(axes=ax)
+        blank_map.draw_limb(axes=ax, color="k")
+        blank_map.draw_grid(axes=ax, color="k")
+
+        return fig, ax
+
+    def plot_map(self, show=True, save=None, **kwargs):
         """
         Plot map, intended as a background for plotting SPICE FOVs
 
@@ -170,6 +197,8 @@ class FovBackground:
             Show figure. Then returns None, None
         save: str
             File to save figure to
+        kwargs: dict
+            Additional parameters for the plotting functions
 
         Return
         ------
@@ -179,12 +208,12 @@ class FovBackground:
             Axes (with relevant projection)
         """
         plot_function = getattr(self, self.map_types[self.map_type])
-        fig, ax = plot_function()  # no need for self anywhere?
+        fig, ax = plot_function(**kwargs)
         return _show_or_save(fig, ax, show, save)
 
 
 def plot_fov_background(
-    map_type=None, cat=None, time=None, observer=None, show=True, save=None
+    map_type=None, cat=None, time=None, observer=None, show=True, save=None, **kwargs
 ):
     """
     Plot map, intended as a background for plotting SPICE FOVs
@@ -203,6 +232,8 @@ def plot_fov_background(
         Show figure. Then returns None, None
     save: str
         File to save figure to
+    kwargs: dict
+        Additional parameters for the plotting functions
 
     Return
     ------
@@ -214,7 +245,7 @@ def plot_fov_background(
     fov_background = FovBackground(
         map_type=map_type, cat=cat, time=time, observer=observer
     )
-    fig, ax = fov_background.plot_map(show=False)
+    fig, ax = fov_background.plot_map(show=False, **kwargs)
     return _show_or_save(fig, ax, show, save)
 
 
@@ -227,6 +258,7 @@ def plot_fovs_with_background(
     fig=None,
     ax=None,
     save=None,
+    bg_kwargs=dict(),
     **kwargs,
 ):
     """
@@ -250,6 +282,8 @@ def plot_fovs_with_background(
         Show figure. Then returns None, None
     save: str
         File to save figure to
+    bg_kwargs: dict
+        Keyword arguments for plot_fov_background()
     kwargs: dict
         Keyword arguments for Catalog.plot_fov()
 
@@ -262,7 +296,12 @@ def plot_fovs_with_background(
     """
     if fig is None or ax is None:
         fig, ax = plot_fov_background(
-            map_type=map_type, cat=cat, time=time, observer=observer, show=False
+            map_type=map_type,
+            cat=cat,
+            time=time,
+            observer=observer,
+            show=False,
+            **bg_kwargs,
         )
     cat.plot_fov(ax, **kwargs)
     return _show_or_save(fig, ax, show, save)